refactor: change the term realtime to time-based where applicable
