+++ src/locale/he.json
@@ -3,8 +3,8 @@
   "dashboard_page_description": "תצוגה גרפית של אחוז הנסיעות שיצאו בפועל מתוך סך הנסיעות המתוכננות, מפולח לפי חברות תחבורה ציבורית (אגד, דן, מטרופולין וכו')",
   "timeline_page_title": "היסטוריית נסיעות",
   "timeline_page_description": "תצוגה של זמני עצירה בתחנות לפי חברה מפעילה, מספר קו ותחנה. השוואה בין שעת העצירה בפועל לעומת שעת העצירה המתוכננת",
-  "realtime_map_page_title": "מפה לפי זמן",
-  "realtime_map_page_description": "תצוגה של מיקום כל האוטובוסים בתאריך, שעה וטווח הדקות שהוכנסו ע\"י המשתמש",
+  "time_based_map_page_title": "מפה לפי זמן",
+  "time_based_map_page_description": "תצוגה של מיקום כל האוטובוסים בתאריך, שעה וטווח הדקות שהוכנסו ע\"י המשתמש",
   "gaps_page_title": "נסיעות שלא בוצעו",
   "gaps_page_description": "תצוגה של כל הנסיעות שלא יצאו כמתוכנן לפי חברה מפעילה, קו, מסלול ותאריך",
   "gaps_patterns_page_title": "דפוסי נסיעות שלא בוצעו",
