+++ src/routes/index.tsx
@@ -6,7 +6,7 @@ const DashboardPage = lazy(() => import('../pages/dashboard/DashboardPage'))
 const TimelinePage = lazy(() => import('../pages/historicTimeline'))
 const GapsPage = lazy(() => import('../pages/gaps'))
 const GapsPatternsPage = lazy(() => import('../pages/gapsPatterns'))
-const RealtimeMapPage = lazy(() => import('../pages/realtimeMap'))
+const TimeBasedMapPage = lazy(() => import('../pages/timeBasedMap'))
 const SingleLineMapPage = lazy(() => import('../pages/singleLineMap'))
 const About = lazy(() => import('../pages/about'))
 const Profile = lazy(() => import('../pages/Profile'))
@@ -61,10 +61,10 @@ export const PAGES = [
     element: <GapsPatternsPage />,
   },
   {
-    label: 'realtime_map_page_title',
+    label: 'time_based_map_page_title',
     path: '/map',
     icon: <HeatMapOutlined />,
-    element: <RealtimeMapPage />,
+    element: <TimeBasedMapPage />,
   },
   {
     label: 'singleline_map_page_title',
